Renamed style and attribute, removed unused progressTipWidth attribute.

+++ animatedprogressbar/src/main/java/com/mckrpk/animatedprogressbar/Utils.kt
@@ -48,6 +48,6 @@ fun getThemePrimaryDarkColor(context: Context, @ColorRes fallbackColorRes: Int):
     }
 }
 
-fun dpToPx(dp: Float, context: Context): Float {
+fun dpToPx(dp: Int, context: Context): Float {
     return dp * (context.resources.displayMetrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)
 }
